+++ Adapter/Adapter.h
@@ -14,7 +14,7 @@ public:
     explicit Adapter(std::shared_ptr<IDatabase> t_db);
     ~Adapter();
 
-    const std::vector<int> GetResult(int) const;
+    const std::vector<size_t> GetResult(int) const;
 
 private:
     Adapter(const Adapter &a) = delete;
